requirements.txt
# --- Core ---
streamlit==1.38.0
python-dotenv==1.0.1

# --- LangChain ecosystem ---
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.29
langchain-openai==0.3.2   # OpenRouter/OpenAI LLM 쓸 때 필요

# --- Search Tools ---
duckduckgo-search==6.2.10   # DuckDuckGo API wrapper (ddgs 기반)

# --- Pydantic v2 (LangChain requires v2) ---
pydantic==2.11.9
pydantic-core==2.33.2

# --- Optional (to avoid warnings) ---
tenacity==8.5.0
SQLAlchemy==2.0.43
requests==2.32.5
httpx==0.28.1
